Rename test file and add some unit tests for Utils.isTxParams
